// packages/shared/src/generated/inputTypeSchemas/CategoryTranslationMinOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const CategoryTranslationMinOrderByAggregateInputSchema: z.ZodType<Prisma.CategoryTranslationMinOrderByAggregateInput> = z.object({
  id: z.lazy(() => SortOrderSchema).optional(),
  categoryId: z.lazy(() => SortOrderSchema).optional(),
  locale: z.lazy(() => SortOrderSchema).optional(),
  name: z.lazy(() => SortOrderSchema).optional(),
}).strict();

export default CategoryTranslationMinOrderByAggregateInputSchema;